Fix clang-tidy `readability-redundant-casting` warnings.

+++ hilti/runtime/src/fiber.cc
@@ -294,7 +294,7 @@ size_t detail::StackBuffer::liveRemainingSize() const {
 #endif
 }
 
-size_t detail::StackBuffer::activeSize() const { return static_cast<size_t>(::fiber_stack_used_size(_fiber)); }
+size_t detail::StackBuffer::activeSize() const { return ::fiber_stack_used_size(_fiber); }
 
 void detail::StackBuffer::save() {
     auto want_buffer_size = std::max(activeSize(), configuration::get().fiber_shared_stack_swap_size_min);
